requirements.txt
annotated-types==0.7.0
anyio==4.13.0
blinker==1.9.0
certifi==2026.2.25
click==8.3.2
distro==1.9.0
eval_type_backport==0.3.1
Flask==3.1.3
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.14.0
jsonpath-python==1.1.5
lxml==6.0.4
MarkupSafe==3.0.3
mistralai==2.3.2
openai==2.31.0
opentelemetry-api==1.39.1
opentelemetry-semantic-conventions==0.60b1
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.8
zipp==3.23.0
gunicorn
